Started to implement the on_initialize function
Started to implement the on_initialize function: initialize all the variable to a zero matrix.
Also declared the functions needed for the real time execution, but have to be implemented.
Divided the computation of the damping and natural frequency matrix into two functions.
Some minor change in the design of the code (with comment and stuff)