Modify integration strategy. Working with small gain (e.g. 1).
Now the change of the position z computed as dt * z \dot is directly added to the current Reference pose of the leg, instead of integrate it (z_k = z_k-1 * dt * z \dot) and then adding z_k to the current reference position.

Also the change of position has been divided in half, so the displacement of one leg is equally distributed between the two legs.